Took logic out of app, moved to controller
